Improve add semester by adding it to NavBar instead of floating randomly, change icon
